refactor: trouble 도메인 정리
