+++ package.json
@@ -41,6 +41,7 @@
     "@types/styled-components": "5.1.34",
     "@vitejs/plugin-react": "^4.4.1",
     "autoprefixer": "^10.4.21",
+    "esbuild": "^0.25.10",
     "eslint": "^9.25.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
