+++ src/entities/trouble/index.ts
@@ -1,0 +1,5 @@
+export * from "./model";
+export * from "./api";
+export { default as TroublogCard } from "./ui/TroublogCard";
+export { default as TroubleShootingCard } from "./ui/TroubleShootingCard";
+export * from "./lib/combinedRoute";
