+++ src/entities/trouble/ui/CardFooterInfo.tsx
