+++ src/entities/trouble/ui/CardTitleSection.tsx
