+++ src/entities/trouble/ui/StatusDot.tsx
