+++ src/pages/Search/SearchResultPage.tsx
@@ -1,4 +1,4 @@
-import TroubleShootingCard from "@/features/mypage/ui/TroubleShootingCard";
+import TroubleShootingCard from "@/entities/trouble/ui/TroubleShootingCard";
 import { PATH } from "@/shared/config/paths";
 import { useInfiniteCommunityTroubleSearch } from "@/hooks/useInfiniteCommunityTroubleSearch";
 import { useInfiniteMyTroubleSearch } from "@/features/mypage/useInfiniteMyTroubleSearch";
@@ -6,7 +6,7 @@ import { useInfiniteUserTroubleSearch } from "@/hooks/useInfiniteUserTroubleSear
 import { useEffect, useRef } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 import { useViewerId } from "@/store/auth";
-import { decideCombined } from "@/utils/combinedRoute";
+import { decideCombined } from "@/entities/trouble/lib/combinedRoute";
 
 // 상태 정규화: API/한글/대소문자 뒤섞여도 'inProgress' | 'complete' | 'created'로 통일
 const normStatus = (raw: any): "inProgress" | "complete" | "created" => {
